the wrong U8model asset/SZU0526_U8input_512x640_preprocess.onnx can run but without meanful output and the theroritally correct asset/SZU0526_floatinput_512x640_preprocess.onnx just migraphx runtime error

+++ detect/ONNX/ONNX.hpp
@@ -35,7 +35,7 @@ private:
     migraphx::program net;
     
     // 修改：输入数据现在是 uint8 类型
-    std::vector<uint8_t> inputTensorValues;
+    std::vector<float> inputTensorValues;
 };
 
 #endif /* _ONNXMODULE_HPP_ */
